Fix invokeMethod did not work
The metatype for uint64_t was not set, which resulted in invokeMethod
silently ignore to call the slots on OBS 29.1.3 on CentOS Stream 9.

+++ src/sync-test-dock.hpp
@@ -28,5 +28,5 @@ private:
 
 private slots:
 	void on_video_marker_found(uint64_t timestamp, double score);
-	void on_audio_marker_found(size_t channel, uint64_t timestamp, double score);
+	void on_audio_marker_found(int channel, uint64_t timestamp, double score);
 };
